% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportedPackages.R
\name{supported.packages}
\alias{supported.packages}
\title{Supported packages}
\usage{
supported.packages()
}
\value{
A character vector
}
\description{
Get a list of currently supported packages
}
\details{
Numerous R-packages are supported, each providing a couple of theoretical statistical distributions for discrete or continuous data. Beside ordinary distributions like normal, t, exponential, ..., some packages implement more exotic distributions like truncrated alpha.
}
\note{
Some of the distributions are redundant, i.e. they are implemented in more than one package.
}
\examples{
sp <- supported.packages()
head(sp)


}
\author{
Markus Boenn
}
